Remove NativeValue realm persistence, and need for ExtendedObject.
There are ways to support NativeValue persistence -- and this was considered
for std.Coins support in contracts; and this could work, but the edge cases to
solve are not easy.  Namely, realm logic should be able to deduce that
references to fields within native values should not be saved as objects
independent of the parent native object.  But beyond that, how can Gno in
general know without fully tracing all internal fields?  Without specialized
"reflect" support or other means, this approach doesn't seem to be solveable to
completion.  And while imperfect support for NativeValue persistence may be
achievable (and already has), it's probably better to avoid the resulting
confusions altgoether.

Instead, we will focus on pure Gno persistence (with GoNativeFunc native
functions (and methods)), and convert std.Coins etc to pure Gno. This requires
building out more standard library functions, so it will happen gradually.