fix(bft/rpc): check http status code (#350)
This change improves the error handling of rpc clients.

Previously, the client tried to unmarshal the response body whatever the
returned status code. This typically outputs errors like `invalid character '<'
looking for beginning of value` because the body format is HTML and not
JSON, and this kind of error is difficult to interpret for the end user.

Now, if the status code is not between 200 and 299, the client doesn't
try to unmarshal the response body and directly returns an error `server at
'<ADDRESS>' returned <ERROR_STATUS>`.