chore: disable auto precompile on gnodev test
we should find a reliable way to re-enable it because it brings more useful
warnings for contracts developers.

but the current implementation is not reliable because it won't automatically
precompile imported packages.

and if the imported package is withing ~/go/mod/pkg, it's read-only. So we
also need to make it working in memory.

Signed-off-by: Manfred Touron <94029+moul@users.noreply.github.com>