chore(website): improve md rendering
Current md rendering involves a small blinking because the md is
displayed first in a <pre> tag, and then it is replaced by the html
transformed by the marked.js library. The change ensures the <pre> tag
is hidden by default, which removes the blinking effect.

To be compliant with browser that doesn't have javascript enabled, a
<noscript> tag has also been added that override the <pre> tag display.
So when javascript is not enabled, the <pre> is visible.

Because the mailchimp form was inside the <pre> tag, it was not rendered
as a form when JS was disabled, only the html tags were visible. To fix
that I moved the mailchimp form outside the <pre>.