chore: attempt to improve OriginCall funcs
`AssertOriginCall` and `IsOriginCall` relies on the number of frames to
determine if the function was called by a gno transaction (definition
taken from `examples/gno.land/r/demo/banktest/README.md`). Since this
number is different in the context of test and filetest, the 2 functions
are overrided during test setup, but the override works only for
filetest files and not for test files.

This change makes `AssertOriginCall` and `IsOriginCall` to rely on
message's signers instead of frames. If the message has at east one
signer, we consider the function was called by a gno tx. As a result,
the functions no longer needs to be overrided in test setup, and they
work both for filetest files and test files.

Fix #481