feat(gnolang): switch fallthrough
Part of #492

Bring support for switch fallthrough. This instruction transfers control
to the next case, even through this case doesn't match.

When executing a fallthrough, you have to remember what is the current
clause index, in order to execute the next one (index+1). For that I used
the switch statement attributes, so I created a new name of attribute
specifically for that. When a switch clause is executed, the clause index
is registered in these attributes. Then, if a fallthrough is met, the
attribute is read to determine what is the next clause to execute.
There's maybe a better way than using attributes, please say so if it's
the case.

The same set of rules as the go-lang has been applied:
- must be in a switch (challenges/switch8.gno)
- mustn't be in a type switch (challenges/switch9.gno)
- mustn't be in the last clause (challenges/switch8b.gno)
- must be the last instruction of the clause (challenges/switch8c.gno)

All switch cases listed in `tests/challenges` are now working, maybe they
should be moved into `tests/files` or `tests/files2` ?