chore(gnolang): revamp switch fallthrough
Found out there's an other `TODO CHALLENGE switch fallthrough` in
pre-process, which makes me realize the proper way of dealing with jump
statements like fallthrough.

In the [initial version][1], a new attribute `ATTR_SWITCH_CLAUSE_IDX` was
introduced to store the clause index, but it appears this kind of thing
should be handled in the pre-process step.

As a result the attribute is removed and replaced with some code in the
pre-process step, that stores the clause index in node.BodyIndex, only in
case we're dealing with a fallthrough.

The tests remain unchanged, which proves this still works as expected.

[1]: https://github.com/gnolang/gno/pull/504