chore: show logger messages in tests/ when -verbose with gnodev test
`f.logger(<...>)` calls in tests/file.go were not shown when `gnodev test -verbose (...)`.
If the cmd user is specifying -verbose (and since we don't have -debug ATM) those traces should be shown

For instance

    if f.logger != nil {
            store.Print()
            f.logger("========================================")
            f.logger("PREPROCESS ALL FILES")
            f.logger("========================================")
    }

Note: I am unsure if store.Print() in above is printed, when -verbose.
this is a different matter