chore(test): support `*OriginCall` funcs in `_test` files
Fix #481

`AssertOriginCall` and `IsOriginCall` functions can be used inside
contracts to determine if the call comes directly from a tx (which is
an *origin call*) or from an other contract (not an *origin call*).

To do this those functions count the number of frames. In production
environment, if there's exactly 2 frames, it is considered as an *origin
call*. In test environment, the number of frames was 3, but it was
working only for `_filetest`, and not for `_test`, where the setup is
different and so the number of frames (7 instead of 3).

To support `_test`, the change brings a way to identify if a test is
a `_test` or `_filetest`, by checking the function name of the first
frame.

There's currently a pending discussion (#683) about how we should deal
with frames, that's why I tried to keep this change minimal, but at
least it fixes something and we can move forward.

Ideally, `*OriginCall` functions should rely on something stronger that
the number of frames, because for instance they could consider that a call
is not an *origin call*, just because `std.IsOriginCall` is called by a
sub-function of the same contract. But that's something we'll probably
tackle with #683.