refactor(tm2/pkg/db): DB interface uses errors.
Refactoring of the DB interface and all its implementations to return
errors.

Panic is not designed for error handling. Panic is designed for handling
exceptional situations, like out-of-bounds array indexing. Using panic
for error handling, we are swallowing some errors and we might cause
inconsistent behavior not cleaning resources properly.

Signed-off-by: Antonio Navarro Perez <antnavper@gmail.com>