feat(gnoland): handle chain panics
Close #756

By turning m.Exception into a slice, it became possible to append
exceptions when panics are chained.

`recover` is adapted so it receives only the last panic, just as in go (see
`tests/files/recover7.gno`). The remaining exception are then wiped.

If no recover is met, the exceptions slice is formatted like in go, with
exception concatened with `\n\t` (see `tests/files/panic0b.gno` that was
prevouisly in `tests/challenges`).