chore: Make applications go-installable
Having a replace expression on a go.mod file makes packages
un-installable using go install command. Removing that replacement. I
checked the fork and there is no extra code in there. If we don't want
to use the latest changes on the original repo, we can stick to an older
version.

Signed-off-by: Antonio Navarro Perez <antnavper@gmail.com>