fix(gnovm): PrevRealm in _test files
When _test files are executed, the pkg path is set to the real path to
the test, which starts by `./examples/gno.land`, this leads to
interpreting the test package as a non-realm, because a realm must
start with `gno.lang/r`.

Because of that, when calling `PrevRealm`, a test package is never
considered as a realm whereas sometime it should.

The fix truncates the `./examples` prefix from the test package, which
ensures it is well detected as a realm when required.

This had several consequences on tests execution, like `testing` and the
test template must not declare global variable, or else the VM tries to
persist them at the end of transaction and that triggers error.