chore(cmd): print shortHelp if longHelp is empty
Currenlty, `shortHelp` is used when the command is listed from its
parent command, and `longHelp` is used when the command is executed with
its help flag.

What I found a little annoying is if `longHelp` is not provided, which
is the case for the majority of commands, there's no text at all when
the command is invoked with the help flag.

This change ensures the `shortHelp` is printed if `longHelp` is empty.