test(gnovm): "*_filetest.gno" instructions
Add tests for Output, Error and Realm instructions

The change introduces a new test dependency
github.com/rogpeppe/go-internal/testscripts, which allows to create
tests bases on txtar files. This is very handy for this kind of tests,
because:
- you can easily creates a set of files that can be used for a command
  (here the `gno test` command)
- you can assert stderr and stdout expectations in a very readable and
  maintainable way
- you can compare files using the `cmp` command, which is very useful to
  assert of the `-update-golden-tests` behavior.

These tests were added to help work on #1015, which will potentially
alter the behavior of "*filetest.gno" instructions updates.