feat: add realm /r/demo/math_eval (#965)
This eval realm is capable of evaluating 32-bit integer expressions as
they would appear in Go. For example, you can use this package (and
realm) to evaluate an expression like `"(32*32+4)-1"` when presented as
a string.

This could have several applications. This current "eval" realm is
simply demonstrating it as a 32-bit integer calculator.

I would like to also add floats, but it seems `strconv.ParseFloat` is
not available yet.

In the future I would like to use the int32 version of this package in
the bytebeat realm (https://github.com/gnolang/gno/pull/840) which could
then utilize this library to generate audio directly from string
expressions, instead of programmed realms - which ultimately gets closer
to an audio NFT platform where expressions are uploaded if they are
parsed correctly.

To test, you can just add the `eval` package and realm:

```
> gnokey maketx addpkg --pkgpath "gno.land/p/demo/math_eval_int32" \
    --pkgdir "examples/gno.land/p/demo/math_eval/int32" \
    --deposit 100000000ugnot --gas-fee 1000000ugnot \
    --gas-wanted 2000000 --broadcast --chainid dev --remote localhost:26657 \
    YOURKEY
> gnokey maketx addpkg --pkgpath "gno.land/r/demo/math_eval" \
    --pkgdir "examples/gno.land/r/demo/math_eval" \
    --deposit 100000000ugnot --gas-fee 1000000ugnot \
    --gas-wanted 2000000 --broadcast --chainid dev --remote localhost:26657 \
    YOURKEY
```

Then open up to `localhost:8888/r/demo/math_eval` and try expressions in
the URL query, like `localhost:8888/r/demo/math_eval:1+1`.



![image](https://github.com/gnolang/gno/assets/6550035/5b227a22-2ce5-4217-9b9b-99c967c9af86)


<details><summary>Checklists...</summary>

## Contributors Checklist

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [x] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](../.benchmarks/README.md).

## Maintainers Checklist

- [ ] Checked that the author followed the guidelines in
`CONTRIBUTING.md`
- [ ] Checked the conventional-commit (especially PR title and verb,
presence of `BREAKING CHANGE:` in the body)
- [ ] Ensured that this PR is not a significant change or confirmed that
the review/consideration process was appropriate for the change
</details>

---------

Co-authored-by: Manfred Touron <94029+moul@users.noreply.github.com>