wip(amino): demo bug with TypeInfo.String
Add a test that demonstrates a bug when the underlying type represented
by TypeInfo has fields of the same type: TypeInfo.String() triggers a
stack overflow because the codec reuses the same TypeInfo instance when
affecting the master TypeInfo.Fields.

This happens when cdc.fullnameToTypeInfo is printed (in case of error
for instance see codec.go:535), causing all registered TypeInfo to be
printed too, which calls their String method.

Example of type that can triggers this issue:
tm2/pkg/crypto/merkle.SimpleProofNode.