fix(codecov): merge coverage for a single upload by workflow (#1190)
It appears that the retry system introduced in #1186 was not sufficient,
as we still hit API limits when running multiple checks concurrently.
This PR merges all coverage files into a single upload at the of each
testing workflows. As a result, we now have 3 uploads instead of 16,
which should drastically reduce the number of Codecov upload failures.

Note: It still appears to fail randomly, but we might need to wait some
time until our API rate decreases. I believe it's still preferable to
have 3 uploads rather than 16, which seemed to overwhelm Codecov.

---------

Signed-off-by: gfanton <8671905+gfanton@users.noreply.github.com>