refactor!: remove `gno build` command (#1297)
Closes: #1242

In favor of `gno precompile -gobuild`. 

As seen in the tests written for `gno build` in this PR #1103, the
command's behavior is kinda unexpected. It actually doesn't care about
the syntax of gno files and only relies on go files found in the path
passed as an argument.

Because of that, I chose to name the flag `-gobuild` instead of `-build`
because I found that represents better what's really happening when this
flag is provided.

As mentioned in the linked issue, `gno build` will be re-implemented in
the future with an other behavior.

<!-- please provide a detailed description of the changes made in this
pull request. -->

<details><summary>Contributors' checklist...</summary>

- [ ] Added new tests, or not needed, or not feasible
- [ ] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [ ] Updated the official documentation or not needed
- [ ] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [ ] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>