docs: reorg & rename top level docs (#1455)
## Description

This PR introduces a small reorg to the docs structure. 

Our docs follow the [Diataxis](https://diataxis.fr/) framework, but upon
further thought I don't believe that we need to or that we should follow
the framework 1:1, as it can be too rigid at times. IMHO it is more
understandable to have the "Explanation" category renamed to "Concepts"
or "Gno Concepts". While this is a minor change, I believe people who
are not aware of Diataxis will find "Concepts" more intuitive.

This PR also extracts the "Gno Tooling" out of "Explanation", as the
category deserves its own spot in the sidebar.

EDIT: This PR also resolves the conflict with and modifies the
`docs/reference/standard-library.md` file (merged with
[#859](https://github.com/gnolang/gno/pull/859)).

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [x] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>